backend/requirements.txt
flask
flask-cors
werkzeug
cloudinary  
python-dotenv 
gunicorn
Pillow